requirements.txt
requests>=2.31.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tabulate>=0.9.0
pytest>=7.4.0
